split core.py into files for designated classes
