+++ devDockerfile
@@ -33,11 +33,16 @@ CMD ls
 COPY api.py /api
 COPY static /api/static
 COPY apidoc.py /api
-COPY core.py  /api
+COPY util.py /api
 COPY sparql.py /api
-COPY pd_stardog_queries.py /api
 COPY corpora.py /api
 COPY corpus.py /api
+COPY poem.py /api
+COPY author.py /api
+COPY pd_stardog_queries.py /api
+COPY pd_corpora.py /api
+COPY pd_corpus.py /api
+COPY pd_poem.py /api
 
 # configure the container to run in an executed manner
 ENTRYPOINT [ "python", "/api/api.py" ]
